+++ pluginlvox/tools/lvox_computedensitythread.h
@@ -8,7 +8,7 @@ class LVOX_ComputeDensityThread : public CT_MonitoredQThread
 {
     Q_OBJECT
 public:
-    LVOX_ComputeDensityThread(CT_Grid3D<double> *densityGrid,
+    LVOX_ComputeDensityThread(CT_Grid3D<float> *densityGrid,
                               CT_Grid3D<int> *hitsGrid,
                               CT_Grid3D<int> *theoriticalGrid,
                               CT_Grid3D<int> *beforeGrid,
@@ -17,7 +17,7 @@ public:
     void run();
 
 private:
-    CT_Grid3D<double>*  _densityGrid;
+    CT_Grid3D<float>*  _densityGrid;
     CT_Grid3D<int>*     _hitsGrid;
     CT_Grid3D<int>*     _theoriticalGrid;
     CT_Grid3D<int>*     _beforeGrid;
