+++ pluginlvox/tools/lvox_distancevisitor.h
@@ -8,12 +8,12 @@ class LVOX_DistanceVisitor : public CT_AbstractGrid3DBeamVisitor
 {
 public:
 
-    LVOX_DistanceVisitor(CT_Grid3D<double> *grid);
+    LVOX_DistanceVisitor(CT_Grid3D<float> *grid);
 
     virtual void visit(const size_t &index, const CT_Beam *beam);
 
 private:
-    CT_Grid3D<double>*  _grid;
+    CT_Grid3D<float>*  _grid;
 };
 
 #endif // LVOX_DISTANCEVISITOR_H
